#3119:added deletion confirmation of the application

+++ apps/pc_frontend/modules/application/actions/actions.class.php
@@ -238,8 +238,12 @@ class applicationActions extends sfActions
       return sfView::ERROR;
     }
 
-    $member_app->delete();
-    return $this->redirect('application/list');
+    if ($request->isMethod(sfRequest::POST))
+    {
+      $member_app->delete();
+      return $this->redirect('application/list');
+    }
+    return sfView::SUCCESS;
   }
 
  /**
